refactor MessageBoxes with dependency injection
