+++ src/year1/pages/Art.tsx
@@ -7,6 +7,8 @@ import Weewa from '../assets/Animations/Elira-Animation_500px.gif'
 import RyuguardA from '../assets/Animations/Finana-AnimationA_500px.gif'
 import RyuguardB from '../assets/Animations/Finana-AnimationB_500px.gif'
 import Pomudachi from '../assets/Animations/Pomu-Animation_500px.gif'
+import ImportMessages from '../components/ImportMessages'
+import MessageBoxesWithCorners from '../components/MessageBoxesWithCorners'
 
 interface Props {
   page: string
@@ -29,17 +31,19 @@ export default function Art (props: Props): JSX.Element {
 
       {/* Art */}
       <MessageBoxesLayout
-        enableLazulight
         titleType='all'
-        enableElira
-        enablePomu
-        enableFinana
-        enableDcl={false}
-        enableFolding={false}
-        enableArt
         heading='All Fan Art'
         page={props.page}
-      />
+      >
+        <MessageBoxesWithCorners messages={ImportMessages({
+          enableLazulight: true,
+          enableElira: true,
+          enablePomu: true,
+          enableFinana: true,
+          enableArt: true
+        })}
+        />
+      </MessageBoxesLayout>
     </>
   )
 }
