+++ src/year1/pages/Messages.tsx
@@ -6,6 +6,8 @@ import Weewa from '../assets/Animations/Elira-Animation_500px.gif'
 import RyuguardA from '../assets/Animations/Finana-AnimationA_500px.gif'
 import RyuguardB from '../assets/Animations/Finana-AnimationB_500px.gif'
 import Pomudachi from '../assets/Animations/Pomu-Animation_500px.gif'
+import ImportMessages from '../components/ImportMessages'
+import MessageBoxesWithCorners from '../components/MessageBoxesWithCorners'
 
 interface Props {
   page: string
@@ -24,16 +26,18 @@ export default function Messages (props: Props): JSX.Element {
       {/* Messages */}
       <MessageBoxesLayout
         titleType='all'
-        enableLazulight
-        enableElira
-        enablePomu
-        enableFinana
-        enableDcl
-        enableFolding={false}
-        enableArt={false}
         heading='All Fan Messages'
         page={props.page}
-      />
+      >
+        <MessageBoxesWithCorners messages={ImportMessages({
+          enableLazulight: true,
+          enableElira: true,
+          enablePomu: true,
+          enableFinana: true,
+          enableDcl: true
+        })}
+        />
+      </MessageBoxesLayout>
     </>
   )
 }
